// src/pdfs/ma/cjd400-motion-to-impound/schema.ts
/** Auto-generated from cjd400-motion-to-impound.pdf — do not edit */
import {
  PDFCheckBox,
  PDFRadioGroup,
  PDFSignature,
  PDFTextField,
} from "@cantoo/pdf-lib";

export const pdfSchema = {
  "Docket No": PDFTextField,
  DefendantRespondent: PDFTextField,
  "Signature of attorney or plaintiff if pro se": PDFSignature,
  BBO: PDFTextField,
  "Docket No_2": PDFTextField,
  "name of party or attorney of record": PDFTextField,
  Address: PDFTextField,
  CityTown_2: PDFTextField,
  State_2: PDFTextField,
  Zip_2: PDFTextField,
  "date of delivery": PDFTextField,
  By: PDFRadioGroup,
  time: PDFTextField,
  "mailing postage paid on": PDFCheckBox,
  "date of mailing": PDFTextField,
  signature: PDFSignature,
  petitionerName: PDFTextField,
  date: PDFTextField,
  residenceStreetAddress: PDFTextField,
  residenceCity: PDFTextField,
  residenceState: PDFTextField,
  residenceZip: PDFTextField,
  phoneNumber: PDFTextField,
  printName: PDFTextField,
  division: PDFTextField,
  motionFor: PDFTextField,
  petitioner: PDFCheckBox,
  division2: PDFTextField,
  request: PDFTextField,
  currentLegalName: PDFTextField,
  motionFor2: PDFTextField,
  motionForPageTwo: PDFTextField,
  dated: PDFTextField,
} as const;

export type PdfFieldName = keyof typeof pdfSchema;